SMS-3534: Adapt PHY Tester Tool application to last changes in PLC PHY Driver and regenerate projects

+++ apps/phy_apps/phy_tester_tool/firmware/src/config/wbz451_curiosity_pl460/system/time/sys_time_definitions.h
@@ -75,9 +75,9 @@ typedef void (*SYS_TIME_PLIB_START)(void);
 typedef void (*SYS_TIME_PLIB_STOP)(void);
 
 
-typedef void (*SYS_TIME_PLIB_PERIOD_SET)(uint16_t period);
-typedef void (*SYS_TIME_PLIB_COMPARE_SET) (uint16_t compare);
-typedef uint16_t (*SYS_TIME_PLIB_COUNTER_GET)(void);
+typedef void (*SYS_TIME_PLIB_PERIOD_SET)(uint32_t period);
+typedef void (*SYS_TIME_PLIB_COMPARE_SET) (uint32_t compare);
+typedef uint32_t (*SYS_TIME_PLIB_COUNTER_GET)(void);
 
 typedef struct
 {
